\encoding{UTF-8}
\name{ProbSamplingI-package}
\alias{ProbSamplingI-package}
\docType{package}

\title{Design and Sampling Strategies for Parameter Estimation and Sample Size Determination.}
\description{
This package provides functions for selecting a sample and estimating parameters such as total, mean, proportion and ratio; through the main sampling designs.
}
\details{
\packageIndices{ProbSamplingI}
Application of probabilistic sampling
}
\author{ Jorge Alberto Barón Cárdenas <jorgeabaron@correo.unicordoba.edu.co>

Guillermo Martínez Flórez <guillermomartinez@correo.unicordoba.edu.co>

Maintainer: Jorge Alberto Barón Cárdenas <jorgeabaron@correo.unicordoba.edu.co>}
\references{
Särndal, C. E., J. H. Wretman, and C. M. Cassel (1992). Foundations of Inference in Survey Sampling. Wiley New York.

Cochran, W. G. (1977). Sampling Techniques, 3ra ed. New York: Wiley.

Thompson, S. K. (1945). Wiley Series in Probability and Statistics, Sampling, 1ra ed. United States of America.
}
\keyword{package}
